Session storage

+++ frontend/src/components/authentication/Login.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Button, FormControl, Input, InputGroup, InputRightElement, VStack, useToast, RadioGroup, Radio, Stack } from '@chakra-ui/react'
+import { Button, FormControl, Input, InputGroup, InputRightElement, VStack, useToast, RadioGroup, Radio, Stack, Checkbox } from '@chakra-ui/react'
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 
@@ -11,6 +11,7 @@ const Login = () => {
   const [username, setUsername] = React.useState()
   const [password, setPassword] = React.useState()
   const [name, setName] = React.useState()
+  const [keepAccess, setKeepAccess] = React.useState(false)
   const [loading, setLoading] = React.useState(false)
 
   const toast = useToast()
@@ -18,8 +19,10 @@ const Login = () => {
 
   const handleClick = () => setShow(!show)
 
+  const handleAccess = () => setKeepAccess(!keepAccess)
+
   const submitUser = async () => {
-    const data = await axios.get('/list-users')
+    const data = await axios.get('/all-users')
     const users = data.data
     const usernames = users.map(item => item.username)
     const emails = users.map(item => item.email)
@@ -75,8 +78,13 @@ const Login = () => {
         duration: 5000,
         isClosable: true,
         position: "bottom",
-      })  
-      localStorage.setItem('userInfo', JSON.stringify(data))
+      })
+      if (keepAccess) {
+        localStorage.setItem('userInfo', JSON.stringify(data))
+        sessionStorage.setItem('userInfo', JSON.stringify(data))
+      } else {
+        sessionStorage.setItem('userInfo', JSON.stringify(data))
+      }
       setLoading(false)
       navigate('/')
     } catch (error) {
@@ -93,7 +101,7 @@ const Login = () => {
   }
 
   const submitCorporate = async () => {
-    const data = await axios.get('/list-corporates')
+    const data = await axios.get('/all-users')
     const corporates = data.data
     const names = corporates.map(item => item.name)
     const emails = corporates.map(item => item.email)
@@ -149,8 +157,13 @@ const Login = () => {
         duration: 5000,
         isClosable: true,
         position: "bottom",
-      })  
-      localStorage.setItem('userInfo', JSON.stringify(data))
+      })
+      if (keepAccess) {
+        localStorage.setItem('userInfo', JSON.stringify(data))
+        sessionStorage.setItem('userInfo', JSON.stringify(data))
+      } else {
+        sessionStorage.setItem('userInfo', JSON.stringify(data))
+      }
       setLoading(false)
       navigate('/')
     } catch (error) {
@@ -183,6 +196,9 @@ const Login = () => {
           <Stack direction='row'>
             <Radio value='1'>User</Radio>
             <Radio value='2'>Corporate</Radio>
+            <Checkbox
+            onChange={handleAccess}
+            >Keep access</Checkbox>
           </Stack>
         </RadioGroup>
       </FormControl>
